Professional approach to persistent settings
Now we have a signleton settings manager that persists through levels.

I did this now because there was an issue I saw, when the user would change the volume, it would set the camera's volume to the new volume - the problem was this would not persist onto the next scene. So I have laid the framework for easy settings implementation. To change the volume via code, do it like this:
SettingsManager.Instance.Volume = floatValueHere

The changeVolume method is pretty much useless I just thought it could be helpful for dynamic float changes.